Fact : 로또 머신의 인터페이스화와 구현
